requirements.txt
Pillow==8.3.2
selenium==4.6.0
webdriver_manager==3.5.2
pytesseract==0.3.8
proxybroker==0.3.2